fix: 修正 long long -> size_t 隐式转换 warning #524

+++ llbc/src/core/log/LogTrace.cpp
@@ -140,7 +140,7 @@ size_t LLBC_LogTraceMgr::GetLogTraceTimes(const LLBC_LogTrace &logTrace) const
         return item.first == logTrace.traceContent;
     });
 
-    return contentIt != traceContents.end() ? contentIt->second : 0;
+    return contentIt != traceContents.end() ? static_cast<size_t>(contentIt->second) : 0;
 }
 
 void LLBC_LogTraceMgr::ClearAllLogTraces()
